refactor(builtins): split startup.rb / builtins.rb into per-class files
startup.rb (2654 lines) now keeps only the RbConfig / load-path
bootstrap code and an explicit require_relative orchestration list;
builtins.rb (845 lines) is removed entirely. Every class/module
fragment moves to its own file under monoruby/builtins/:

- new files: basic_object, kernel (all four Kernel fragments plus
  caller_locations and the format/sprintf alias, merged in original
  order), object, class, regexp, exception (NoMatchingPattern
  errors + Exception + FloatDomainError), module, warning, process,
  file (File::Constants + File constants + the spec-fix class
  methods), fiber, thread (Thread, Mutex/Queue/SizedQueue/
  ConditionVariable, ThreadGroup, top-level aliases), boolean
  (TrueClass/FalseClass/NilClass), marshal, gc, object_space, io,
  encoding, io_buffer, data, enumerator, hash, file_stat, proc,
  method, dir, time, match_data, filetest, env, argf
- appended to existing per-class files: array.rb (values_at),
  float/symbol/rational/set/struct.rb (the ruby/spec
  instance_method-identity alias fixups, placed after the methods
  they re-point)

Load order is preserved exactly: the pre-comparable zone never
references Comparable/Enumerable at load time (ObjectSpace::WeakMap
still skips its `include ::Enumerable`), IO::Buffer loads after
comparable, the include-Enumerable files load after enumerable, and
ARGF stays last. Rust-side comments that pointed at builtins.rb now
point at the split files.

Verified: a line-multiset comparison against the original two files
shows no lost content (only the intentional comment replacement and
the removed `require_relative 'builtins'`); build succeeds and a
smoke script covering the moved surfaces (alias identity, Data,
Queue, File::Stat, WeakMap) matches the pre-split behavior.
