+++ monoruby/builtins/basic_object.rb
@@ -1,0 +1,11 @@
+class BasicObject
+  private
+  def singleton_method_added(name)
+  end
+
+  def singleton_method_removed(name)
+  end
+
+  def singleton_method_undefined(name)
+  end
+end
