+++ monoruby/builtins/float.rb
@@ -57,3 +57,14 @@ class Float
     self / other
   end
 end
+
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Float
+  alias === ==
+  alias modulo %
+  alias quo fdiv
+end
