+++ monoruby/builtins/marshal.rb
@@ -1,0 +1,4 @@
+module Marshal
+  MAJOR_VERSION = 4
+  MINOR_VERSION = 8
+end
