+++ monoruby/builtins/match_data.rb
@@ -1,0 +1,8 @@
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class MatchData
+  alias deconstruct captures
+end
