+++ monoruby/builtins/rational.rb
@@ -130,3 +130,12 @@ class Rational
     [numerator, denominator]
   end
 end
+
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Rational
+  alias magnitude abs
+end
