+++ monoruby/builtins/regexp.rb
@@ -1,0 +1,7 @@
+class Regexp
+  IGNORECASE = 1
+  EXTENDED = 2
+  MULTILINE = 4
+  FIXEDENCODING = 16
+  NOENCODING = 32
+end
