+++ monoruby/builtins/set.rb
@@ -33,3 +33,15 @@ module Enumerable
     klass.new(self, *args, &block)
   end
 end
+
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Set
+  alias < proper_subset?
+  alias > proper_superset?
+  alias << add
+  alias eql? ==
+end
