+++ monoruby/builtins/startup.rb
@@ -179,2074 +179,35 @@ RbConfig::CONFIG['sitearch']   = RUBY_PLATFORM
 RbConfig::CONFIG['target']     = "#{__host_cpu}-pc-#{__host_os}"
 RbConfig::CONFIG['host']       = "#{__host_cpu}-pc-#{__host_os}"
 
-class BasicObject
-  private
-  def singleton_method_added(name)
-  end
-
-  def singleton_method_removed(name)
-  end
-
-  def singleton_method_undefined(name)
-  end
-end
-
-module Kernel
-  private
-
-  # Internal helper: coerce value to Integer via to_int.
-  # Raises TypeError with CRuby-compatible message if conversion fails.
-  def __to_int(val)
-    return val if val.is_a?(Integer)
-    if val.respond_to?(:to_int)
-      result = val.to_int
-      unless result.is_a?(Integer)
-        raise TypeError, "can't convert #{val.class} into Integer (#{val.class}#to_int gives #{result.class})"
-      end
-      result
-    else
-      raise TypeError, "no implicit conversion of #{val.class} into Integer"
-    end
-  end
-
-  # Internal helper: coerce value to String via to_str.
-  # Raises TypeError with CRuby-compatible message if conversion fails.
-  def __to_str(val)
-    return val if val.is_a?(String)
-    if val.respond_to?(:to_str)
-      result = val.to_str
-      unless result.is_a?(String)
-        raise TypeError, "can't convert #{val.class} into String (#{val.class}#to_str gives #{result.class})"
-      end
-      result
-    else
-      raise TypeError, "no implicit conversion of #{val.class} into String"
-    end
-  end
-end
-
-class Object
-  # NOTE: no `def initialize` here — the default constructor is the
-  # native `BasicObject#initialize` (arity 0), matching CRuby's owner
-  # and its strict arity (`Object.new(1)` raises ArgumentError). A
-  # Ruby-level `def initialize(...)` would silently accept any
-  # arguments and also defeat the JIT's forwarding specialization for
-  # argument-less `Class#new`.
-
-  def tap
-    yield self
-    self
-  end
-
-  def itself
-    self
-  end
-
-  def then
-    return to_enum(:then) { 1 } unless block_given?
-    yield self
-  end
-  alias yield_self then
-
-  def respond_to_missing?(method_name, include_private = false)
-    false
-  end
-
-  def <=>(other)
-    return 0 if equal?(other)
-    # The `self == other` fallback would recurse infinitely when
-    # `==` is `Comparable#==` (a Comparable class that did not
-    # override `==`): that `==` calls `<=>`, which reaches here
-    # again. Skip the fallback in that case (covers both a missing
-    # `<=>` and a user `<=>` that calls `super`). Net behaviour
-    # matches CRuby, which uses an equivalent recursion guard.
-    return nil if self.class.instance_method(:==).owner == Comparable
-    0 if (self == other)
-  end
-end
-
-class Class
-  # The canonical object constructor, written in Ruby. `__builtin_allocate__`
-  # is the private, non-overridable allocator (users override `allocate`,
-  # which `new` deliberately bypasses, as in CRuby); `__builtin_initialize__`
-  # is the privileged spelling that dispatches to `initialize` bypassing its
-  # `private` visibility. The `(...)` forward is optimized end-to-end: the
-  # interpreter defers the rest-Array via the lazy-forwarding marker, and a
-  # specialized JIT compile source-routes the caller's argument slots
-  # straight into `initialize`'s frame (D1), so construction allocates
-  # nothing but the object itself.
-  def new(...)
-    o = __builtin_allocate__
-    o.__builtin_initialize__(...)
-    o
-  end
-
-  private
-  def inherited(subclass)
-  end
-end
-
-class Regexp
-  IGNORECASE = 1
-  EXTENDED = 2
-  MULTILINE = 4
-  FIXEDENCODING = 16
-  NOENCODING = 32
-end
-
-# Raised when a `case/in` exhausts its branches or an `expr => pattern`
-# match fails (the pattern-matching desugar references these by name).
-class NoMatchingPatternError < StandardError
-end
-
-class NoMatchingPatternKeyError < NoMatchingPatternError
-  def initialize(message = nil, matchee: nil, key: nil)
-    @matchee = matchee
-    @key = key
-    super(message)
-  end
-
-  attr_reader :matchee, :key
-end
-
-class Module
-  # Partial ordering on the include / inherit graph, matching CRuby's
-  # `rb_mod_cmp`:
-  #
-  # | relationship                                    | result |
-  # | ----------------------------------------------- | ------ |
-  # | `self` is `other`                               | `0`    |
-  # | `self` is a subclass of, or includes, `other`   | `-1`   |
-  # | `other` is a subclass of, or includes, `self`   | `+1`   |
-  # | unrelated                                       | `nil`  |
-  # | `other` is not a Module / Class                 | `nil`  |
-  #
-  # Defined on Module so `<` / `>` / `<=` / `>=` (provided by
-  # `Comparable` in CRuby; we synthesise them) all reduce to this.
-  def <=>(other)
-    return nil unless other.is_a?(Module)
-    return 0 if equal?(other)
-    if ancestors.include?(other)
-      -1
-    elsif other.ancestors.include?(self)
-      +1
-    else
-      nil
-    end
-  end
-
-  private
-  def extended(mod)
-  end
-
-  def prepended(mod)
-  end
-
-  def included(mod)
-  end
-
-  # CRuby keeps these as private instance methods of Module so subclasses
-  # override them with `def method_added(name); …; end` (i.e. via the
-  # default-private visibility inside a `class M`). `public` here would
-  # surface them as `Module.public_instance_methods` entries, which the
-  # spec explicitly disallows ("is a private instance method").
-  def method_added(name)
-  end
-
-  def method_removed(name)
-  end
-
-  def method_undefined(name)
-  end
-
-  def const_added(name)
-  end
-
-  public
-  def const_missing(name)
-    # Drop the implicit `Object::` prefix so a top-level miss reads
-    # `uninitialized constant Foo` (not `Object::Foo`), matching CRuby.
-    # For a qualified miss, CRuby crafts the prefix from the module's
-    # `#name`, falling back to `#inspect` when the module is anonymous.
-    if self.equal?(Object)
-      qual = name.to_s
-    else
-      prefix = self.name || self.inspect
-      qual = "#{prefix}::#{name}"
-    end
-    raise NameError.new("uninitialized constant #{qual}", name)
-  end
-
-  def include?(mod)
-    # CRuby `Module#include?` accepts only true Modules, not Classes (even
-    # though `Class < Module`). And the receiver is *not* counted as one of
-    # its own included modules — `M.include?(M)` is `false`.
-    if !mod.is_a?(Module) || mod.is_a?(Class)
-      raise TypeError, "wrong argument type #{mod.class} (expected Module)"
-    end
-    return false if equal?(mod)
-    ancestors.include?(mod)
-  end
-
-  # `Module.used_refinements` returns the refinements active in the
-  # current scope. monoruby has no refinement support, so this returns
-  # an empty Array as a permissive mock — gems and code that
-  # defensively read this list (RSpec, Sorbet) won't crash. Defined in
-  # Ruby (not Rust) so the user can override it in specs that actually
-  # exercise refinements. CRuby only exposes the class form (no
-  # `Module#used_refinements` instance method), so we follow suit.
-  def self.used_refinements
-    []
-  end
-end
-
-module Warning
-  @categories = {
-    deprecated: false, experimental: true, performance: false,
-    strict_unused_block: false,
-  }
-
-  def self.categories
-    @categories.keys
-  end
-
-  def self.[](category)
-    unless category.is_a?(Symbol)
-      raise TypeError, "wrong argument type #{category.class} (expected Symbol)"
-    end
-    unless @categories.key?(category)
-      raise ArgumentError, "unknown category: #{category}"
-    end
-    @categories[category]
-  end
-
-  def self.[]=(category, value)
-    unless category.is_a?(Symbol)
-      raise TypeError, "wrong argument type #{category.class} (expected Symbol)"
-    end
-    unless @categories.key?(category)
-      raise ArgumentError, "unknown category: #{category}"
-    end
-    @categories[category] = value ? true : false
-  end
-
-  def warn(msg, category: nil)
-    $stderr.write(msg)
-    nil
-  end
-  extend self
-end
-
-module Kernel
-  # Faithful Kernel#warn: combine the (flattened) messages, gate by
-  # $VERBOSE / Warning[category], then delegate to Warning.warn so a
-  # user-overridden Warning.warn takes effect (matches CRuby).
-  def warn(*messages, uplevel: nil, category: nil)
-    messages = messages.flatten
-    return nil if messages.empty?
-
-    unless category.nil?
-      if category.is_a?(Symbol)
-        # already a Symbol
-      elsif category.respond_to?(:to_sym)
-        category = category.to_sym
-        unless category.is_a?(Symbol)
-          raise TypeError, "can't convert to Symbol"
-        end
-      else
-        raise TypeError,
-              "no implicit conversion of #{category.class} into Symbol"
-      end
-    end
-
-    # $VERBOSE == nil silences all warnings.
-    return nil if $VERBOSE.nil?
-
-    # Category gating: outside verbose mode a known but disabled
-    # category suppresses the message; an unknown category is passed
-    # through.
-    if category && $VERBOSE != true
-      enabled =
-        begin
-          Warning[category]
-        rescue ArgumentError
-          true
-        end
-      return nil unless enabled
-    end
-
-    str = +""
-    unless uplevel.nil?
-      uplevel = __to_int(uplevel)
-      raise ArgumentError, "negative level (#{uplevel})" if uplevel < 0
-      loc = caller_locations(uplevel + 1, 1)
-      if loc && (loc = loc.first)
-        str << "#{loc.path}:#{loc.lineno}: warning: "
-      end
-    end
-    messages.each do |m|
-      s = m.to_s
-      str << s
-      str << "\n" unless s.end_with?("\n")
-    end
-
-    # Pass the category keyword, but fall back to a positional-only
-    # call if Warning.warn was redefined without keyword support
-    # (matches CRuby, and stays robust when Warning.warn is a mock).
-    begin
-      Warning.warn(str, category: category)
-    rescue ArgumentError
-      Warning.warn(str)
-    end
-    nil
-  end
-  module_function :warn
-
-  # Internal: deprecation warnings raised from the Rust runtime (e.g.
-  # assigning non-nil to $/). Routed through Kernel#warn so the
-  # :deprecated category gating and Warning.warn overrides apply.
-  def __warn_deprecated(msg)
-    warn(msg, category: :deprecated)
-  end
-  module_function :__warn_deprecated
-
-  # Internal: the prism lowerer desugars a top-level `return <arg>`
-  # into a call to this helper followed by the return, so the warning
-  # fires only when the return actually executes (CRuby behavior).
-  # rb_warn-level: silenced by -W0 ($VERBOSE == nil), not by default.
-  def __warn_toplevel_return(path)
-    warn("#{path}: warning: argument of top-level return is ignored") unless $VERBOSE.nil?
-  end
-  module_function :__warn_toplevel_return
-end
-
-module Process
-  CLOCK_REALTIME = 0
-  CLOCK_MONOTONIC = 1
-  CLOCK_PROCESS_CPUTIME_ID = 2
-  CLOCK_THREAD_CPUTIME_ID	= 3
-  CLOCK_MONOTONIC_RAW	= 4
-  CLOCK_REALTIME_COARSE	= 5
-  CLOCK_MONOTONIC_COARSE = 6
-  CLOCK_BOOTTIME = 7
-  CLOCK_REALTIME_ALARM = 8
-  CLOCK_BOOTTIME_ALARM = 9
-
-  # Reap child `pid` and return a thread whose #value waits for it and yields
-  # the resulting Process::Status. monoruby's Thread is cooperative (the block
-  # runs at #value/#join time), which matches Open3's "drain the pipes first,
-  # then read the exit status" ordering.
-  def self.detach(pid)
-    Thread::Waiter.new(pid)
-  end
-
-  class Tms
-    attr_accessor :utime, :stime, :cutime, :cstime
-  end
-
-  # Wraps the raw POSIX wait(2) status word (as returned by `waitpid`).
-  # Callers pass the raw int and pid; all predicates decode the bit layout:
-  #   bits 0..6  = termination signal (0 = normal exit)
-  #   bit  7     = core dumped flag
-  #   bits 8..15 = exit code (for normal exit)
-  #   low byte 0x7F = stopped (SIGSTOP etc.)
-  class Status
-    attr_reader :pid
-
-    def initialize(raw_status, pid)
-      @status = raw_status
-      @pid = pid
-    end
-
-    def exited?
-      (@status & 0x7F) == 0
-    end
-
-    def exitstatus
-      exited? ? (@status >> 8) & 0xFF : nil
-    end
-
-    def signaled?
-      low = @status & 0x7F
-      low != 0 && low != 0x7F
-    end
-
-    def termsig
-      signaled? ? @status & 0x7F : nil
-    end
-
-    def stopped?
-      (@status & 0xFF) == 0x7F
-    end
-
-    def stopsig
-      stopped? ? (@status >> 8) & 0xFF : nil
-    end
-
-    def coredump?
-      (@status & 0x80) != 0
-    end
-
-    def success?
-      exited? ? exitstatus == 0 : nil
-    end
-
-    def to_i
-      @status
-    end
-
-    def to_s
-      if signaled?
-        "pid #{@pid} SIG#{termsig}#{coredump? ? ' (core dumped)' : ''}"
-      elsif stopped?
-        "pid #{@pid} stopped SIG#{stopsig}"
-      else
-        "pid #{@pid} exit #{exitstatus}"
-      end
-    end
-
-    def inspect
-      "#<Process::Status: #{to_s}>"
-    end
-
-    def ==(other)
-      if other.is_a?(Integer)
-        to_i == other
-      else
-        super
-      end
-    end
-  end
-end
-
-module File::Constants
-  FNM_SYSCASE = 0
-  FNM_NOESCAPE = 1
-  FNM_PATHNAME = 2
-  FNM_DOTMATCH = 4
-  FNM_CASEFOLD = 8
-  FNM_EXTGLOB = 16
-  # Open(2) flags and flock(2) operations (Linux values), mirrored on
-  # File itself below via `include File::Constants` in CRuby; monoruby
-  # historically defined them directly on File, so keep both in sync.
-  RDONLY   = 0
-  WRONLY   = 1
-  RDWR     = 2
-  APPEND   = 1024
-  CREAT    = 64
-  EXCL     = 128
-  TRUNC    = 512
-  NOCTTY   = 256
-  NONBLOCK = 2048
-  DSYNC    = 4096
-  SYNC     = 1052672
-  RSYNC    = 1052672
-  DIRECT   = 16384
-  NOFOLLOW = 131072
-  # No-op outside Windows; defined for source compatibility (CRuby).
-  SHARE_DELETE = 0
-  BINARY   = 0
-  LOCK_SH  = 1
-  LOCK_EX  = 2
-  LOCK_UN  = 8
-  LOCK_NB  = 4
-  NULL = "/dev/null"
-end
-
-class File
-  include File::Constants
-  FNM_SYSCASE = 0
-  FNM_NOESCAPE = 1
-  FNM_PATHNAME = 2
-  FNM_DOTMATCH = 4
-  FNM_CASEFOLD = 8
-  FNM_EXTGLOB = 16
-  NULL = "/dev/null"
-  BINARY = 0
-
-  RDONLY   = 0
-  WRONLY   = 1
-  RDWR     = 2
-  APPEND   = 1024
-  CREAT    = 64
-  EXCL     = 128
-  TRUNC    = 512
-  NOCTTY   = 256
-  NONBLOCK = 2048
-  DSYNC    = 4096
-  SYNC     = 1052672
-  RSYNC    = 1052672
-  DIRECT   = 16384
-  NOFOLLOW = 131072
-  SHARE_DELETE = 0
-  LOCK_SH  = 1
-  LOCK_EX  = 2
-  LOCK_UN  = 8
-  LOCK_NB  = 4
-
-  Separator = "/"
-  SEPARATOR = "/"
-  ALT_SEPARATOR = nil
-  PATH_SEPARATOR = ":"
-end
-
-class Fiber
-  @@main_fiber = nil
-  def self.current
-    # Inside a fiber body the executor knows which Fiber it is running as;
-    # at a thread's root context there is none, so fall back to a shared
-    # main-fiber object (distinct from any explicit Fiber, which is all that
-    # per-Fiber mutex ownership needs — cross-thread checks key on the
-    # Thread).
-    __current_fiber || (@@main_fiber ||= Fiber.new {})
-  end
-
-  # --- Fiber-local storage (CRuby 3.2+) --------------------------------
-  #
-  # `Fiber.current.storage` is a Hash whose keys must be Symbols. The
-  # `Fiber.[key]` / `Fiber.[]=key, val` class-method sugar reads/writes
-  # the *current* fiber's storage. Pure-Ruby implementation here —
-  # monoruby tracks an `@storage` ivar per Fiber, lazily materialised.
-
-  # Returns the storage Hash. CRuby returns `nil` until `[]=` /
-  # `storage=` materialises one (no lazy init on bare read), so do
-  # not auto-create here.
-  def storage
-    @storage
-  end
-
-  # Assigns the storage Hash. `nil` clears it. Non-Hash raises
-  # TypeError; non-Symbol keys raise TypeError.
-  def storage=(hash)
-    if hash.nil?
-      @storage = nil
-      return nil
-    end
-    raise TypeError, "no implicit conversion of #{hash.class} into Hash" \
-      unless hash.is_a?(Hash)
-    hash.each_key do |k|
-      raise TypeError, "wrong argument type #{k.class} (expected Symbol)" \
-        unless k.is_a?(Symbol)
-    end
-    @storage = hash
-  end
-
-  # Class-method sugar over `Fiber.current` storage. Symbol-keyed.
-  # `Fiber[:k]` on a fiber with no storage returns nil (CRuby).
-  def self.[](key)
-    raise TypeError, "wrong argument type #{key.class} (expected Symbol)" \
-      unless key.is_a?(Symbol)
-    s = current.storage
-    s ? s[key] : nil
-  end
-
-  # `Fiber[:k] = v` materialises an empty storage on first write.
-  def self.[]=(key, value)
-    raise TypeError, "wrong argument type #{key.class} (expected Symbol)" \
-      unless key.is_a?(Symbol)
-    cur = current
-    cur.storage = {} if cur.storage.nil?
-    cur.storage[key] = value
-  end
-
-  # --- Scheduler -------------------------------------------------------
-  #
-  # monoruby has no fiber scheduler, so `set_scheduler` is effectively
-  # a no-op store. CRuby allows `nil` to clear the scheduler, so accept
-  # that too. `scheduler` returns whatever was set (default nil).
-
-  @@scheduler = nil
-
-  def self.scheduler
-    @@scheduler
-  end
-
-  def self.set_scheduler(scheduler)
-    @@scheduler = scheduler
-  end
-
-
-  def self.current_scheduler
-    # CRuby: nil when the current fiber is blocking. monoruby fibers
-    # are always blocking, so this is always nil.
-    nil
-  end
-
-  # --- Blocking / non-blocking ----------------------------------------
-  #
-  # CRuby contracts (3.2+):
-  #   `Fiber.blocking?`         returns `1` when the *current* fiber
-  #                             is blocking, `false` otherwise.
-  #   `Fiber#blocking?`         instance form: `true` / `false`.
-  # monoruby has no non-blocking fibers, so the class form is `1`
-  # and the instance form is `true`.
-
-  def self.blocking?
-    1
-  end
-
-  def blocking?
-    true
-  end
-end
-
-class Thread
-  # monoruby threads are cooperative green threads multiplexed on the one
-  # OS thread by the native scheduler (src/scheduler.rs). The native side
-  # provides: Thread.new / .start / .fork (queue a body), .current /
-  # .main / .list / .pass / .stop, and #join / #value / #status /
-  # #alive? / #stop? / #wakeup / #run. Blocking APIs park the calling
-  # thread on the scheduler; a body starts running the first time any
-  # thread reaches a blocking point.
-  #
-  # This Ruby side keeps only the pure-bookkeeping surface: a name,
-  # thread-/fiber-local storage, and an interrupt-mask no-op (real
-  # asynchronous interruption — #raise / #kill — is not implemented yet,
-  # so there is nothing to mask).
-
-  def self.each_caller_location
-    caller_locations(1).each do |loc|
-      yield loc
-    end
-  end
-
-  # Whether a detected deadlock is ignored rather than aborting. monoruby
-  # does not run the deadlock detector, but the flag round-trips.
-  @@ignore_deadlock = false
-  def self.ignore_deadlock
-    @@ignore_deadlock
-  end
-
-  def self.ignore_deadlock=(flag)
-    @@ignore_deadlock = flag
-  end
-
-  attr_accessor :name
-
-  # Reported when a thread dies with an unhandled exception (read by the
-  # native finalizer). Falls back to the class-level default
-  # (`Thread.report_on_exception`), which defaults to true, as in CRuby.
-  def report_on_exception
-    @report_on_exception.nil? ? Thread.report_on_exception : @report_on_exception
-  end
-
-  def report_on_exception=(flag)
-    @report_on_exception = flag
-  end
-
-  # The group this thread was added to (ThreadGroup#add), defaulting to
-  # ThreadGroup::Default. The timeout gem consults
-  # `watcher.group.enclosed?` when it spawns its watchdog thread.
-  def group
-    @__thread_group || ThreadGroup::Default
-  end
-
-  # CRuby's Thread#to_s: `#<Thread:0xADDR[@name] status>` (the spawn
-  # file:line is not tracked). ruby/spec interpolates this into the
-  # report_on_exception output matchers, so the `#<Thread:` prefix and
-  # the status word matter.
-  def to_s
-    st = status
-    st = "dead" unless st
-    if (n = @name)
-      "#<Thread:#{format('0x%016x', object_id << 1)}@#{n} #{st}>"
-    else
-      "#<Thread:#{format('0x%016x', object_id << 1)} #{st}>"
-    end
-  end
-  alias inspect to_s
-
-  # When set (per-thread, or the Thread global default), an exception
-  # that terminates this thread is re-raised in the main thread (read by
-  # the native finalizer, which also suppresses the report in that case,
-  # as in CRuby). Default false.
-  def abort_on_exception
-    @abort_on_exception.nil? ? false : @abort_on_exception
-  end
-
-  def abort_on_exception=(flag)
-    @abort_on_exception = flag
-  end
-
-  class << self
-    def abort_on_exception
-      @abort_on_exception.nil? ? false : @abort_on_exception
-    end
-
-    def abort_on_exception=(flag)
-      @abort_on_exception = flag
-    end
-
-    # Global default for Thread#report_on_exception (true, as in CRuby).
-    # Threads whose instance flag is unset fall back to this; ruby/spec's
-    # spec_helper requires the class-level accessor to exist (otherwise
-    # it installs a shim instance method that raises).
-    def report_on_exception
-      @report_on_exception.nil? ? true : @report_on_exception
-    end
-
-    def report_on_exception=(flag)
-      @report_on_exception = flag
-    end
-  end
-
-  # Thread objects are native (ObjTy::THREAD) and no longer run a Ruby
-  # initialize, so the local-storage tables are created lazily.
-
-  # Normalize a fiber-local / thread-variable key to a Symbol, as CRuby
-  # does: Symbols pass through, Strings convert, an object with #to_str
-  # converts through that, anything else is a TypeError.
-  def __thread_key(key)
-    case key
-    when Symbol
-      key
-    when String
-      key.to_sym
-    else
-      if key.respond_to?(:to_str)
-        key.to_str.to_sym
-      else
-        raise TypeError, "#{key.inspect} is not a symbol nor a string"
-      end
-    end
-  end
-  private :__thread_key
-
-  def [](key)
-    k = __thread_key(key)
-    @fiber_locals && @fiber_locals[k]
-  end
-
-  def []=(key, value)
-    raise FrozenError, "can't modify frozen thread locals" if frozen?
-    (@fiber_locals ||= {})[__thread_key(key)] = value
-  end
-
-  def key?(key)
-    k = __thread_key(key)
-    !!(@fiber_locals && @fiber_locals.key?(k))
-  end
-
-  def keys
-    @fiber_locals ? @fiber_locals.keys : []
-  end
-
-  def fetch(key, *default)
-    if default.size > 1
-      raise ArgumentError, "wrong number of arguments (given #{default.size + 1}, expected 1..2)"
-    end
-    k = __thread_key(key)
-    if @fiber_locals && @fiber_locals.key?(k)
-      @fiber_locals[k]
-    elsif block_given?
-      warn "warning: block supersedes default value argument" unless default.empty?
-      yield(key)
-    elsif !default.empty?
-      default[0]
-    else
-      raise KeyError.new("key not found: #{key.inspect}", receiver: self, key: key)
-    end
-  end
-
-  def thread_variable_get(key)
-    k = __thread_key(key)
-    @thread_variables && @thread_variables[k]
-  end
-
-  def thread_variable_set(key, value)
-    raise FrozenError, "can't modify frozen thread locals" if frozen?
-    (@thread_variables ||= {})[__thread_key(key)] = value
-  end
-
-  def thread_variable?(key)
-    k = __thread_key(key)
-    !!(@thread_variables && @thread_variables.key?(k))
-  end
-
-  def thread_variables
-    @thread_variables ? @thread_variables.keys : []
-  end
-
-  # monoruby has no real thread-scheduler priority; store the value so it
-  # round-trips (a new thread defaults to 0, matching an unset priority).
-  def priority
-    @priority || 0
-  end
-
-  def priority=(value)
-    unless value.is_a?(Integer)
-      unless value.respond_to?(:to_int) && (value = value.to_int).is_a?(Integer)
-        raise TypeError, "no implicit conversion of #{value.class} into Integer"
-      end
-    end
-    # CRuby clamps the stored priority to -3..3.
-    value = 3 if value > 3
-    value = -3 if value < -3
-    @priority = value
-  end
-
-  # A per-Thread identifier while the thread is alive (nil once it has
-  # finished). monoruby multiplexes green threads onto one OS thread, so
-  # this is a distinct-per-object token rather than a real kernel tid.
-  def native_thread_id
-    alive? ? object_id : nil
-  end
-
-  # The object returned by `Process.detach`. Reaping one specific child via
-  # `Process.wait2` is a *terminating* operation (unlike an arbitrary thread
-  # body), so — unlike a general Thread, which defines no #join / #value —
-  # the waiter safely runs the reaper on #join / #value. A missing child
-  # (`Errno::ECHILD`) yields nil, matching CRuby. This is what keeps Open3's
-  # `wait_thr.value` / `wait_thr.join` working.
-  class Waiter < Thread
-    # The native Thread.new requires a block (it queues a green thread);
-    # a Waiter is an inert shell around one specific child pid, so build
-    # it via allocate.
-    def self.new(pid)
-      w = allocate
-      w.__send__(:__init_waiter, pid)
-      w
-    end
-
-    def __init_waiter(pid)
-      @pid = pid
-      self[:pid] = pid
-    end
-    private :__init_waiter
-
-    attr_reader :pid
-
-    def value
-      __reap
-    end
-
-    def join(limit = nil)
-      __reap
-      self
-    end
-
-    private
-
-    def __reap
-      return @status if @reaped
-      @reaped = true
-      @status = begin
-        Process.wait2(@pid)[1]
-      rescue SystemCallError
-        nil
-      end
-    end
-  end
-
-  # With timeslice preemption, a switch can occur at any safepoint —
-  # every method call and loop back-edge. These pure-Ruby primitives
-  # stay correct under that model through two rules:
-  #
-  # 1. Test-and-set sequences contain no safepoint between the test and
-  #    the set: straight-line ivar reads/writes only, with any needed
-  #    method calls (e.g. `Thread.current`) hoisted before the test.
-  #    See `Mutex#try_lock`.
-  # 2. Every "register as waiter → park" sequence tolerates a wakeup
-  #    landing in the window before the park: `Thread#wakeup` on a
-  #    *running* thread arms its park permit, making the next park
-  #    return immediately (scheduler.rs, `ThreadInner::park_permit`).
-  #    The park sites all sit in retry loops, so a spurious early
-  #    return is re-checked.
-
-  class Mutex
-    def locked?
-      # A thread releases its locks when it terminates (CRuby semantics), so
-      # an owner that is no longer alive means the lock has been abandoned
-      # and the mutex is effectively unlocked.
-      (o = @owner) ? o.alive? : false
-    end
-
-    def owned?
-      # Ownership is per-Fiber (CRuby): the same Thread's other Fibers do not
-      # own a mutex this Fiber locked. The Thread test also makes a cross-
-      # thread check false regardless of Fiber identity.
-      @owner == Thread.current && @owner_fiber == Fiber.current
-    end
-
-    def try_lock
-      # Atomic under preemption: `Thread.current` is a method call (a
-      # safepoint where a timeslice switch can occur), so it must be
-      # evaluated BEFORE the test. The test-and-set below is straight-line
-      # code — ivar read, branch, ivar write — with no calls and no loop
-      # back-edges, hence no safepoint can interleave another thread
-      # between the test and the set.
-      cur = Thread.current
-      cur_fiber = Fiber.current
-      if @owner
-        false
-      else
-        @owner = cur
-        @owner_fiber = cur_fiber
-        true
-      end
-    end
-
-    def lock
-      # Deadlock detection is per-Thread, not per-Fiber: a mutex already held
-      # by THIS thread (in any fiber) can only be released by this thread, so
-      # trying to acquire it again — recursively or from a sibling fiber —
-      # would park the one thread that could release it. Raise rather than
-      # hang. (Ownership *reported* by #owned? is still per-Fiber.)
-      raise ThreadError, "deadlock; recursive locking" if @owner == Thread.current
-      until try_lock
-        # Reclaim a lock abandoned by a terminated owner (a dead thread
-        # never unlocks or wakes waiters, so a contender must take it over
-        # rather than park forever). Clearing @owner is idempotent under a
-        # race — concurrent reclaimers write the same nil — and the actual
-        # acquisition still goes through try_lock's atomic test-and-set, so
-        # no two threads can both win.
-        if (o = @owner) && !o.alive?
-          @owner = nil
-          next
-        end
-        (@waiters ||= []) << Thread.current
-        begin
-          Thread.stop
-        ensure
-          @waiters.delete(Thread.current)
-        end
-      end
-      self
-    end
-
-    def unlock
-      unless @owner
-        raise ThreadError, "Attempt to unlock a mutex which is not locked"
-      end
-      # Thread-level check (not the per-Fiber #owned?): keeps the hot unlock
-      # path free of the extra Fiber.current safepoint, and every unlock spec
-      # exercises the cross-thread case anyway.
-      unless @owner == Thread.current
-        raise ThreadError, "Attempt to unlock a mutex which is locked by another thread"
-      end
-      @owner = nil
-      # Hand the next waiter a chance; it re-contends in its lock loop.
-      if @waiters
-        while (w = @waiters.shift)
-          if w.alive?
-            # Permit-arming wake: unlike the public Thread#wakeup, a wake
-            # racing the waiter's park in Mutex#lock (preempted between
-            # registering and parking) is not lost — the next park returns
-            # immediately and the lock loop re-contends.
-            w.__wakeup_permit
-            break
-          end
-        end
-      end
-      self
-    end
-
-    def synchronize
-      raise ThreadError, "must be called with a block" unless block_given?
-      lock
-      begin
-        yield
-      ensure
-        unlock
-      end
-    end
-
-    # Atomically release the mutex, sleep, and re-acquire it on wake
-    # (including on an exception raised into the sleeper). Returns the
-    # rounded number of seconds actually slept, as in CRuby. A negative
-    # duration is an ArgumentError (validated before the ownership check).
-    def sleep(timeout = nil)
-      if timeout && timeout < 0
-        raise ArgumentError, "time interval must not be negative"
-      end
-      raise ThreadError, "Attempt to unlock a mutex which is not locked" unless owned?
-      start = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-      unlock
-      begin
-        timeout ? Kernel.sleep(timeout) : Thread.stop
-      ensure
-        lock
-      end
-      (Process.clock_gettime(Process::CLOCK_MONOTONIC) - start).round
-    end
-  end
-
-  class Queue
-    # Layered on Mutex + ConditionVariable, exactly like CRuby's
-    # thread_sync.c. Mutual exclusion (not statement ordering) is what
-    # makes check-and-take atomic here: with callee-entry GC/preempt
-    # polls, EVERY method call is a safepoint, so an unlocked
-    # `@items.empty?` / `@items.shift` pair lets a competing consumer
-    # take the last item in between and the shift returns a phantom nil.
-    #
-    # NOTE: the wait loops are `while true`, never `loop do` —
-    # `Kernel#loop` swallows StopIteration and ClosedQueueError is a
-    # StopIteration subclass (so `loop { q.pop }` ends cleanly on close);
-    # a `raise ClosedQueueError` inside a `loop` block would be silently
-    # eaten and turned into a nil return.
-    def initialize
-      @items = []
-      @closed = false
-      @mutex = Mutex.new
-      @cv_pop = ConditionVariable.new
-    end
-
-    def push(item)
-      @mutex.synchronize do
-        raise ClosedQueueError, "queue closed" if @closed
-        @items.push(item)
-        @cv_pop.signal
-      end
-      self
-    end
-    alias << push
-    alias enq push
-
-    # CRuby-compatible `timeout:` validation, shared by Queue#pop,
-    # SizedQueue#pop and SizedQueue#push. `false` must NOT be treated as
-    # "no timeout": without the type check it fell through to the
-    # unbounded `Thread.stop` branch and blocked forever.
-    def __check_timeout(timeout, non_block)
-      return nil if timeout.nil?
-      raise ArgumentError, "can't set a timeout if non_block is enabled" if non_block
-      return timeout if Numeric === timeout
-      case timeout
-      when true, false, String
-        desc = String === timeout ? "string" : timeout.inspect
-        raise TypeError, "no implicit conversion to float from #{desc}"
-      end
-      unless timeout.respond_to?(:to_f)
-        raise TypeError, "can't convert #{timeout.class} into Float"
-      end
-      f = timeout.to_f
-      unless Float === f
-        raise TypeError,
-              "can't convert #{timeout.class} to Float (#{timeout.class}#to_f gives #{f.class})"
-      end
-      f
-    end
-    private :__check_timeout
-
-    def pop(non_block = false, timeout: nil)
-      timeout = __check_timeout(timeout, non_block)
-      deadline = timeout && Process.clock_gettime(Process::CLOCK_MONOTONIC) + timeout
-      @mutex.synchronize do
-        while true
-          return __pop_locked unless @items.empty?
-          # CRuby raises even on a closed queue: only a *blocking* pop
-          # returns nil for closed-and-drained.
-          raise ThreadError, "queue empty" if non_block
-          return nil if @closed
-          if deadline
-            remaining = deadline - Process.clock_gettime(Process::CLOCK_MONOTONIC)
-            return nil if remaining <= 0
-            @cv_pop.wait(@mutex, remaining)
-          else
-            @cv_pop.wait(@mutex)
-          end
-        end
-      end
-    end
-    alias shift pop
-    alias deq pop
-
-    def empty?
-      @items.empty?
-    end
-
-    def size
-      @items.size
-    end
-    alias length size
-
-    def clear
-      @mutex.synchronize { @items.clear }
-      self
-    end
-
-    # Closing wakes every parked consumer/producer: consumers drain the
-    # remaining items then get nil; producers raise ClosedQueueError.
-    def close
-      @mutex.synchronize do
-        return self if @closed
-        @closed = true
-        @cv_pop.broadcast
-      end
-      self
-    end
-
-    def closed?
-      @closed
-    end
-
-    def num_waiting
-      @cv_pop.__num_waiting
-    end
-
-    private
-
-    # Take the head item; the caller holds @mutex and has checked
-    # non-emptiness. SizedQueue overrides this to also free a producer.
-    def __pop_locked
-      @items.shift
-    end
-  end
-
-  class SizedQueue < Queue
-    def initialize(max)
-      max = max.to_int if !max.is_a?(Integer) && max.respond_to?(:to_int)
-      raise ArgumentError, "queue size must be positive" unless max.is_a?(Integer) && max > 0
-      super()
-      @max = max
-      @cv_push = ConditionVariable.new
-    end
-
-    attr_reader :max
-
-    def max=(new_max)
-      raise ArgumentError, "queue size must be positive" unless new_max.is_a?(Integer) && new_max > 0
-      @mutex.synchronize do
-        grew = new_max > @max
-        @max = new_max
-        # Growing frees capacity: every parked producer re-checks.
-        @cv_push.broadcast if grew
-      end
-      new_max
-    end
-
-    def push(item, non_block = false, timeout: nil)
-      timeout = __check_timeout(timeout, non_block)
-      deadline = timeout && Process.clock_gettime(Process::CLOCK_MONOTONIC) + timeout
-      @mutex.synchronize do
-        while true
-          raise ClosedQueueError, "queue closed" if @closed
-          if @items.size < @max
-            @items.push(item)
-            @cv_pop.signal
-            return self
-          end
-          raise ThreadError, "queue full" if non_block
-          if deadline
-            remaining = deadline - Process.clock_gettime(Process::CLOCK_MONOTONIC)
-            return nil if remaining <= 0
-            @cv_push.wait(@mutex, remaining)
-          else
-            @cv_push.wait(@mutex)
-          end
-        end
-      end
-    end
-    alias << push
-    alias enq push
-
-    def clear
-      @mutex.synchronize do
-        @items.clear
-        # Clearing frees capacity (CRuby wakes blocked producers here).
-        @cv_push.broadcast
-      end
-      self
-    end
-
-    def close
-      @mutex.synchronize do
-        return self if @closed
-        @closed = true
-        @cv_pop.broadcast
-        @cv_push.broadcast
-      end
-      self
-    end
-
-    def num_waiting
-      @cv_pop.__num_waiting + @cv_push.__num_waiting
-    end
-
-    private
-
-    def __pop_locked
-      v = @items.shift
-      # Popping frees a slot: let one parked producer through.
-      @cv_push.signal
-      v
-    end
-  end
-
-  class ConditionVariable
-    def initialize
-      @waiters = []
-    end
-
-    # Atomically release `mutex` and park until signaled (or the timeout
-    # elapses), then re-acquire `mutex` before returning — also on an
-    # exception raised into the waiter.
-    def wait(mutex, timeout = nil)
-      @waiters << Thread.current
-      begin
-        # Delegate to Mutex#sleep so the mutex is atomically released,
-        # the caller parks, and the mutex is re-acquired on wake — exactly
-        # as CRuby does (which is why `#wait` calls `#sleep` on its arg).
-        mutex.sleep(timeout)
-      ensure
-        @waiters.delete(Thread.current)
-      end
-      self
-    end
-
-    # ConditionVariable holds runtime synchronization state that cannot be
-    # serialized; CRuby raises TypeError from Marshal.
-    def marshal_dump
-      raise TypeError, "can't dump #{self.class}"
-    end
-
-    def signal
-      while (w = @waiters.shift)
-        if w.alive?
-          # Permit-arming wake (see Mutex#unlock): a signal racing the
-          # waiter's park in CV#wait must not be lost.
-          w.__wakeup_permit
-          break
-        end
-      end
-      self
-    end
-
-    def broadcast
-      until @waiters.empty?
-        signal
-      end
-      self
-    end
-
-    # Waiter-count peek for Queue#num_waiting (a waiter registers before
-    # releasing the mutex and stays registered until it re-acquires it,
-    # so this tracks "parked in #wait" closely and race-free).
-    def __num_waiting
-      @waiters.size
-    end
-  end
-
-  class Backtrace
-    class Location
-      def initialize(frame)
-        @frame = frame.to_s
-        if @frame =~ /\A(.+):(\d+):in ['`](.+)'\z/
-          @path = $1
-          @lineno = $2.to_i
-          @label = $3
-        else
-          @path = @frame
-          @lineno = 0
-          @label = ""
-        end
-      end
-
-      attr_reader :path, :lineno, :label
-
-      def base_label
-        l = @label
-        l = $1 while l =~ /\Ablock (?:\(\d+ levels\) )?in (.+)\z/
-        l
-      end
-
-      # CRuby returns nil for frames without a real file (eval'd code,
-      # internal frames).
-      def absolute_path
-        return nil if @path.start_with?("(") || @path.start_with?("<")
-        File.expand_path(@path)
-      end
-
-      def to_s
-        @frame
-      end
-      alias inspect to_s
-    end
-  end
-end
-
-# Top-level aliases (CRuby compatibility)
-class ThreadError < StandardError; end unless defined?(::ThreadError)
-# Raised by push/pop on a closed queue (CRuby: subclass of StopIteration,
-# so `loop { q.pop }` exits cleanly when the queue is closed).
-class ClosedQueueError < StopIteration; end unless defined?(::ClosedQueueError)
-Queue = Thread::Queue
-SizedQueue = Thread::SizedQueue
-
-# Thread groups. Threads carry their group in an ivar (nil = Default);
-# scheduling is untouched — groups are pure bookkeeping, which matches
-# their CRuby role. The timeout gem needs `Thread#group`,
-# `ThreadGroup#enclosed?` and `ThreadGroup::Default.add`.
-class ThreadGroup
-  def initialize
-    @enclosed = false
-  end
-
-  def enclose
-    @enclosed = true
-    self
-  end
-
-  def enclosed?
-    @enclosed
-  end
-
-  def add(thread)
-    if @enclosed
-      raise ThreadError, "can't move to the enclosed thread group"
-    end
-    if thread.group.enclosed?
-      raise ThreadError, "can't move from the enclosed thread group"
-    end
-    thread.instance_variable_set(:@__thread_group, self)
-    self
-  end
-
-  def list
-    Thread.list.select { |t| t.group == self }
-  end
-
-  Default = new
-end
-ConditionVariable = Thread::ConditionVariable
-
-class Exception
-  def backtrace_locations
-    # Locations come from the raise-time capture, independent of any
-    # string backtrace installed via #set_backtrace (CRuby: after
-    # `set_backtrace(strings)` on a never-raised exception,
-    # #backtrace_locations stays nil). Memoized so repeated calls
-    # return the same, mutable Array.
-    # A prior #backtrace_locations memo, or locations installed via
-    # set_backtrace(locations), win over re-deriving from the capture.
-    return @backtrace_locations if defined?(@backtrace_locations) && @backtrace_locations
-    frames = __raise_backtrace
-    return nil if frames.nil?
-    @backtrace_locations = frames.map { |f| Thread::Backtrace::Location.new(f) }
-  end
-
-  # CRuby 3.4+: `set_backtrace` (and thus `$@ =`) also accepts an Array
-  # of `Thread::Backtrace::Location`, stored as their string forms. Mixed
-  # or otherwise invalid arrays fall through to the native checker, which
-  # raises the usual TypeError.
-  def set_backtrace(bt)
-    if bt.is_a?(Array) && !bt.empty? && bt.all? { |e| Thread::Backtrace::Location === e }
-      # CRuby 3.4+: an Array of Locations sets both the string backtrace
-      # and #backtrace_locations.
-      __set_backtrace(bt.map(&:to_s))
-      @backtrace_locations = bt
-    else
-      # Setting a string backtrace leaves #backtrace_locations as it was
-      # (the raise-time capture, or nil), so don't touch the memo here.
-      __set_backtrace(bt)
-    end
-  end
-
-  # CRuby's Exception#inspect goes through #to_s (which user classes
-  # may override): `#<ClassName: to_s>`, or just the class name when
-  # #to_s returns an empty string.
-  def inspect
-    s = to_s
-    if s.nil? || s.empty?
-      self.class.name || self.class.to_s
-    else
-      "#<#{self.class}: #{s}>"
-    end
-  end
-
-  # Whether the uncaught-exception report would be written to a tty —
-  # the default for `full_message`'s :highlight option.
-  def self.to_tty?
-    $stderr.equal?(STDERR) && STDERR.tty?
-  rescue NoMethodError
-    false
-  end
-
-  def full_message(highlight: nil, order: :top, **opts)
-    # CRuby's first line is `bt[0]: <detailed_message>` (i.e.
-    # "message (ClassName)"), followed by `\tfrom <frame>` lines.
-    # `--backtrace-limit=N` truncates the from-lines to N entries plus
-    # a `\t ... K levels...` marker, exactly like the top-level
-    # uncaught-error report. All keyword arguments except :order are
-    # forwarded to #detailed_message, with :highlight resolved to its
-    # default first.
-    highlight = Exception.to_tty? if highlight.nil?
-    msg = nil
-    if respond_to?(:detailed_message)
-      dm = detailed_message(highlight: highlight, **opts)
-      dm = dm.to_str if !dm.is_a?(String) && dm.respond_to?(:to_str)
-      msg = dm if dm.is_a?(String)
-    end
-    if msg.nil?
-      # No usable #detailed_message: fall back to the class name.
-      msg = highlight ? "\e[1;4m#{self.class}\e[m" : self.class.to_s
-    end
-    bt = backtrace
-    # An exception that was never raised has no backtrace; CRuby shows
-    # the caller of full_message instead.
-    bt = caller if bt.nil? || bt.empty?
-    if bt && !bt.empty?
-      rest = bt[1..]
-      limit = Kernel.__backtrace_limit
-      trailer = nil
-      if limit && rest.size > limit
-        trailer = "\t ... #{rest.size - limit} levels...\n"
-        rest = rest[0, limit]
-      end
-      out = if order == :bottom
-        header = highlight ? "\e[1mTraceback\e[m (most recent call last):\n"
-                           : "Traceback (most recent call last):\n"
-        # CRuby numbers the reversed frames by their distance from the
-        # error line: `\tN: from <frame>` counting down to 1.
-        numbered = []
-        i = rest.size
-        rest.each do |l|
-          numbered << "\t#{i}: from #{l}\n"
-          i -= 1
-        end
-        numbered << trailer if trailer
-        header + numbered.reverse.join + "#{bt[0]}: #{msg}\n"
-      else
-        lines = rest.map { |l| "\tfrom #{l}\n" }
-        lines << trailer if trailer
-        "#{bt[0]}: #{msg}\n" + lines.join
-      end
-    else
-      out = msg + "\n"
-    end
-    # Chain the cause's report so every exception in the chain appears.
-    c = cause
-    out += c.full_message(highlight: highlight, order: :top) if c
-    out
-  end
-
-  # CRuby `Exception#detailed_message(highlight: false, **)`:
-  # decorate the message with the class name; empty-message and
-  # anonymous-class cases have special forms.
-  def detailed_message(highlight: false, **)
-    msg = message.to_s
-    cls = self.class
-    if msg.empty?
-      base = instance_of?(::RuntimeError) ? "unhandled exception" : (cls.name || cls.to_s)
-      return highlight ? "\e[1;4m#{base}\e[m" : base
-    end
-    first, *rest = msg.split("\n", -1)
-    # With :highlight every line of a multi-line message is bolded
-    # individually (`\e[1m…\e[m` per line).
-    if cls.name.nil?
-      head = highlight ? "\e[1m#{first}\e[m" : first
-    elsif highlight
-      head = "\e[1m#{first} (\e[1;4m#{cls}\e[m\e[1m)\e[m"
-    else
-      head = "#{first} (#{cls})"
-    end
-    if highlight
-      rest = rest.map { |l| l.empty? ? l : "\e[1m#{l}\e[m" }
-    end
-    ([head] + rest).join("\n")
-  end
-end
-
-Mutex = Thread::Mutex
-
-class TrueClass
-  class << self
-    undef_method :new
-  end
-  TRUE_TO_S = "true".freeze
-  def to_s
-    TRUE_TO_S
-  end
-end
-
-class FalseClass
-  class << self
-    undef_method :new
-  end
-  FALSE_TO_S = "false".freeze
-  def to_s
-    FALSE_TO_S
-  end
-end
-
-class NilClass
-  class << self
-    undef_method :new
-  end
-  NIL_TO_S = "".freeze
-  def to_s
-    NIL_TO_S
-  end
-
-  def to_a
-    []
-  end
-
-  def to_i
-    0
-  end
-
-  def to_f
-    0.0
-  end
-
-  def to_h
-    {}
-  end
-
-  def =~(_other)
-    nil
-  end
-
-  def to_c
-    Complex(0, 0)
-  end
-
-  def to_r
-    Rational(0)
-  end
-
-  def rationalize(*args)
-    if args.length > 1
-      raise ArgumentError, "wrong number of arguments (given #{args.length}, expected 0..1)"
-    end
-    Rational(0)
-  end
-end
-
-module Marshal
-  MAJOR_VERSION = 4
-  MINOR_VERSION = 8
-end
-
-module Kernel
-  module_function
-
-  def puts(*args)
-    $stdout.puts(*args)
-    nil
-  end
-
-  # `Kernel#print` is a native builtin (see builtins/kernel.rs): it
-  # delegates to `$stdout.print` but additionally writes `$_` when
-  # called with no arguments, which a pure-Ruby definition can't do
-  # because `$_` is frame-local (it would read print's own nil slot).
-
-  def p(*args)
-    if args.size == 1
-      $stdout.puts(args[0].inspect)
-      args[0]
-    elsif args.empty?
-      nil
-    else
-      args.each { |a| $stdout.puts(a.inspect) }
-      args
-    end
-  end
-
-  def printf(fmt = nil, *args)
-    if fmt.respond_to?(:write)
-      fmt.write(format(*args))
-    elsif fmt
-      $stdout.write(format(fmt, *args))
-    end
-    nil
-  end
-
-  # Prevent CRuby's bundled_gems.rb from patching require with warning
-  # logic.  monoruby provides its own implementations of formerly-bundled
-  # gems (fiddle, strscan, etc.) so the warnings are not applicable.
-  # bundled_gems.rb's replace_require checks for this method and returns
-  # early if it already exists.
-  alias no_warning_require require
-  module_function :no_warning_require
-end
-
-module Kernel
-  module_function
-
-  def open(name, *args, &block)
-    if name.respond_to?(:to_open)
-      name.to_open(*args, &block)
-    else
-      name = name.to_path if name.respond_to?(:to_path)
-      name = name.to_str if name.respond_to?(:to_str)
-      raise TypeError, "no implicit conversion of #{name.class} into String" unless name.is_a?(String)
-      File.open(name, *args, &block)
-    end
-  end
-
-  def String(arg)
-    return arg if arg.is_a?(::String)
-    if arg.respond_to?(:to_str)
-      result = arg.to_str
-      return result if result.is_a?(::String)
-      raise TypeError, "can't convert #{arg.class} to String (#{arg.class}#to_str gives #{result.class})"
-    end
-    result = arg.to_s
-    unless result.is_a?(::String)
-      raise TypeError, "can't convert #{arg.class} to String (#{arg.class}#to_s gives #{result.class})"
-    end
-    result
-  end
-
-  def Hash(arg)
-    return {} if arg.nil?
-    return {} if arg.is_a?(::Array) && arg.empty?
-    return arg if arg.is_a?(::Hash)
-    if arg.respond_to?(:to_hash)
-      result = arg.to_hash
-      return {} if result.nil?
-      return result if result.is_a?(::Hash)
-      raise TypeError, "can't convert #{arg.class} to Hash (#{arg.class}#to_hash gives #{result.class})"
-    end
-    raise TypeError, "can't convert #{arg.class} into Hash"
-  end
-
-  def srand(*args)
-    Random.srand(*args)
-  end
-
-  def putc(ch)
-    if ch.is_a?(String)
-      s = ch[0]
-    else
-      i = ch.is_a?(Integer) ? ch : __to_int(ch)
-      s = (i & 0xff).chr
-    end
-    $stdout.write(s)
-    ch
-  end
-
-  # Kernel#test(cmd, file1[, file2]) — minimal subset of CRuby's file-test
-  # operator. `cmd` accepts either a single-character string or its integer
-  # code-point.
-  def test(cmd, file1, file2 = nil)
-    c = cmd.is_a?(Integer) ? cmd : cmd.to_s.ord
-    case c
-    when ?e.ord then File.exist?(file1)
-    when ?f.ord then File.file?(file1)
-    when ?d.ord then File.directory?(file1)
-    when ?r.ord then File.readable?(file1)
-    when ?R.ord then File.readable_real?(file1)
-    when ?w.ord then File.writable?(file1)
-    when ?W.ord then File.writable_real?(file1)
-    when ?x.ord then File.executable?(file1)
-    when ?X.ord then File.executable_real?(file1)
-    when ?l.ord then File.symlink?(file1)
-    when ?p.ord then File.pipe?(file1)
-    when ?S.ord then File.socket?(file1)
-    when ?b.ord then File.blockdev?(file1)
-    when ?c.ord then File.chardev?(file1)
-    when ?u.ord then File.setuid?(file1)
-    when ?g.ord then File.setgid?(file1)
-    when ?k.ord then File.sticky?(file1)
-    when ?o.ord then File.owned?(file1)
-    when ?G.ord then File.grpowned?(file1)
-    when ?s.ord
-      sz = File.size?(file1)
-      sz && sz > 0 ? sz : nil
-    when ?z.ord then File.zero?(file1)
-    when ?M.ord then File.mtime(file1)
-    when ?A.ord then File.atime(file1)
-    when ?C.ord then File.ctime(file1)
-    when ?-.ord then File.identical?(file1, file2)
-    when ?=.ord then File.mtime(file1) == File.mtime(file2)
-    when ?<.ord then File.mtime(file1) < File.mtime(file2)
-    when ?>.ord then File.mtime(file1) > File.mtime(file2)
-    else
-      raise ArgumentError, "unknown command '#{cmd.is_a?(Integer) ? cmd.chr : cmd}'"
-    end
-  end
-end
-
-module GC
-  # `GC.count` and `GC.start` are implemented in Rust (they read the
-  # allocator's collection counter / force a full collection).
-
-  # --- boolean mode flags -------------------------------------------------
-  # monoruby has no stress mode, no auto-compaction and no time
-  # measurement, but the accessors round-trip a stored value so callers
-  # (and specs) can set and read them back.
-  @stress = false
-  @measure_total_time = false
-  @auto_compact = false
-
-  def self.stress
-    @stress
-  end
-
-  def self.stress=(flag)
-    @stress = flag
-  end
-
-  def self.measure_total_time
-    @measure_total_time
-  end
-
-  def self.measure_total_time=(flag)
-    @measure_total_time = flag
-  end
-
-  def self.auto_compact
-    @auto_compact
-  end
-
-  def self.auto_compact=(flag)
-    @auto_compact = flag
-  end
-
-  # Total time spent in GC, in nanoseconds. monoruby does not measure
-  # this, so report 0 (an Integer, never decreasing).
-  def self.total_time
-    0
-  end
-
-  # --- GC.config ----------------------------------------------------------
-  def self.config(*args)
-    @config ||= { implementation: "default", rgengc_allow_full_mark: true }
-    return @config.dup if args.empty?
-    arg = args[0]
-    return @config.dup if arg.nil?
-    unless arg.is_a?(Hash)
-      raise ArgumentError, "GC.config requires a Hash argument"
-    end
-    # `:implementation` is read-only.
-    if arg.key?(:implementation)
-      raise ArgumentError, 'Attempting to set read-only key "Implementation"'
-    end
-    # Update known knobs (coercing arbitrary truthy/falsey values to a
-    # boolean, as MRI does); unknown keys are ignored.
-    arg.each do |k, v|
-      next unless @config.key?(k)
-      @config[k] = v ? true : false
-    end
-    @config.dup
-  end
-
-  # Instance form (available via `obj.extend(GC)`); always returns nil.
-  def garbage_collect(full_mark: true, immediate_mark: true, immediate_sweep: true)
-    GC.start
-    nil
-  end
-
-  module Profiler
-    @enabled = false
-
-    def self.enabled?
-      @enabled
-    end
-
-    def self.enable
-      @enabled = true
-      nil
-    end
-
-    def self.disable
-      @enabled = false
-      nil
-    end
-
-    def self.clear
-      nil
-    end
-
-    def self.result
-      ""
-    end
-
-    def self.report(out = $stdout)
-      out.print(result)
-      nil
-    end
-
-    def self.total_time
-      0.0
-    end
-  end
-end
-
-# Minimal ObjectSpace stub for monoruby.
-#
-# monoruby has no support for weak references or general object iteration.
-# Provide just enough surface area for libraries that defensively reference
-# ObjectSpace constants (ActiveSupport::DescendantsTracker, ConnectionPool,
-# weakref) — `WeakMap` is implemented as a strong-referenced hash so keys
-# are never GCed while held by the map. That is semantically weaker than
-# CRuby but correct enough for class loading and most non-GC-sensitive use.
-module ObjectSpace
-  class WeakMap
-    include ::Enumerable if defined?(::Enumerable)
-
-    def initialize
-      @map = {}
-    end
-
-    def [](key)
-      @map[key.object_id]&.first
-    end
-
-    def []=(key, value)
-      @map[key.object_id] = [value, key]
-      value
-    end
-
-    def key?(key)
-      @map.key?(key.object_id)
-    end
-    alias include? key?
-    alias member? key?
-
-    def delete(key)
-      pair = @map.delete(key.object_id)
-      pair ? pair.first : nil
-    end
-
-    def keys
-      @map.values.map { |pair| pair[1] }
-    end
-
-    def values
-      @map.values.map { |pair| pair[0] }
-    end
-
-    def each
-      return to_enum(:each) unless block_given?
-      @map.each_value { |pair| yield pair[1], pair[0] }
-      self
-    end
-    alias each_pair each
-
-    def each_key
-      return to_enum(:each_key) unless block_given?
-      @map.each_value { |pair| yield pair[1] }
-      self
-    end
-
-    def each_value
-      return to_enum(:each_value) unless block_given?
-      @map.each_value { |pair| yield pair[0] }
-      self
-    end
-
-    def size
-      @map.size
-    end
-    alias length size
-
-    def inspect
-      "#<ObjectSpace::WeakMap:#{format('0x%016x', object_id << 1)} size=#{size}>"
-    end
-  end
-
-  def self.each_object(klass = nil)
-    return to_enum(:each_object, klass) unless block_given?
-    0
-  end
-
-  # Register a finalizer for +obj+. The finalizer (a callable or block,
-  # invoked with the object's id) is run at program termination. monoruby
-  # never runs finalizers asynchronously at GC time, which the spec
-  # explicitly permits. The actual registry lives in the runtime; the
-  # private +__register_finalizer+ primitive records the pair.
-  def self.define_finalizer(obj, *args, &block)
-    callable = block || args[0]
-    if callable.nil?
-      raise ArgumentError, "wrong number of arguments (given 1, expected 2)"
-    end
-    unless callable.respond_to?(:call)
-      raise ArgumentError, "no _id2ref or finalizer is given; must respond to #call"
-    end
-    # The primitive returns the effective callable: the one already
-    # registered when an equal finalizer was given before, else +callable+.
-    [0, __register_finalizer(obj, callable)]
-  end
-
-  def self.undefine_finalizer(obj)
-    __unregister_finalizer(obj)
-  end
-
-  def self.garbage_collect(**opts)
-    GC.start(**opts)
-  end
-
-  def self._id2ref(id)
-    raise RangeError, "0x#{id.to_s(16)} is not id value"
-  end
-
-  def self.count_objects(result_hash = {})
-    result_hash[:TOTAL] = 0
-    result_hash[:FREE] = 0
-    result_hash
-  end
-end
-
-class IO
-  include File::Constants
-  # Enumerable is mixed in after `require_relative 'enumerable'` at the
-  # bottom of this file (the module is not defined yet at this point).
-
-  SEEK_SET = 0
-  SEEK_CUR = 1
-  SEEK_END = 2
-
-  # Mix-ins tagging the exceptions raised by `*_nonblock` when the
-  # operation would block. The concrete classes subclass the matching
-  # Errno (on Linux EAGAIN == EWOULDBLOCK) and include these so
-  # `rescue IO::WaitReadable` / `rescue IO::WaitWritable` work.
-  module WaitReadable; end
-  module WaitWritable; end
-  class EAGAINWaitReadable < Errno::EAGAIN
-    include IO::WaitReadable
-  end
-  class EAGAINWaitWritable < Errno::EAGAIN
-    include IO::WaitWritable
-  end
-  EWOULDBLOCKWaitReadable = EAGAINWaitReadable
-  EWOULDBLOCKWaitWritable = EAGAINWaitWritable
-  # Raised by Socket#connect_nonblock while the connect is in flight
-  # (`rescue IO::WaitWritable`, then IO.select writable and retry).
-  class EINPROGRESSWaitWritable < Errno::EINPROGRESS
-    include IO::WaitWritable
-  end
-
-  # The sync flag is per-object bookkeeping only: monoruby's writes are
-  # unbuffered, so semantically every IO behaves as if sync were on. The
-  # flag still round-trips through #sync= and defaults to true for the
-  # process's stderr (fd 2), matching CRuby.
-  def sync
-    raise IOError, "closed stream" if closed?
-    s = @sync
-    s.nil? ? fileno == 2 : s
-  end
-
-  def sync=(v)
-    raise IOError, "closed stream" if closed?
-    @sync = v ? true : false
-    v
-  end
-
-  # CRuby's IO#putc: a String argument writes its first character, any
-  # other argument is converted with #to_int and the low byte is written.
-  def putc(ch)
-    if ch.is_a?(String)
-      write(ch[0])
-    else
-      i = ch.is_a?(Integer) ? ch : (ch.respond_to?(:to_int) ? ch.to_int : nil)
-      raise TypeError, "no implicit conversion of #{ch.class} into Integer" if i.nil?
-      write((i & 0xff).chr)
-    end
-    ch
-  end
-
-  # CRuby raises EOFError (not RuntimeError) at end of stream for the
-  # `read*` family; #getbyte/#getc return nil at EOF.
-  def readbyte
-    raise IOError, "closed stream" if closed?
-    b = getbyte
-    raise EOFError, "end of file reached" if b.nil?
-    b
-  end
-
-  def readchar
-    raise IOError, "closed stream" if closed?
-    c = getc
-    raise EOFError, "end of file reached" if c.nil?
-    c
-  end
-
-  def each_byte
-    raise IOError, "closed stream" if closed?
-    return to_enum(:each_byte) { nil } unless block_given?
-    while (b = getbyte)
-      yield b
-    end
-    self
-  end
-
-  def each_char
-    raise IOError, "closed stream" if closed?
-    return to_enum(:each_char) { nil } unless block_given?
-    while (c = getc)
-      yield c
-    end
-    self
-  end
-
-  def each_codepoint
-    raise IOError, "closed stream" if closed?
-    return to_enum(:each_codepoint) { nil } unless block_given?
-    each_char { |c| yield c.ord }
-    self
-  end
-
-  # Enumerator-replay helper: positional-only arguments survive the
-  # to_enum round-trip (a `chomp:` keyword would come back as a
-  # positional Hash and be misparsed as a separator/limit).
-  def __each_line(args, chomp)
-    # #gets implements the full (sep, limit, chomp:) semantics natively.
-    while (line = gets(*args, chomp: chomp))
-      yield line
-    end
-    self
-  end
-  private :__each_line
-
-  def each_line(*args, chomp: false, **)
-    raise IOError, "closed stream" if closed?
-    # CRuby rejects a zero limit up front (a zero-limit #gets would
-    # return "" forever). The limit is the second positional argument,
-    # or the first when it is not a (nil/String) separator.
-    lim = if args.size >= 2
-            args[1]
-          elsif args.size == 1 && !args[0].nil? && !args[0].is_a?(String)
-            args[0]
-          end
-    if lim.is_a?(Integer) && lim == 0
-      raise ArgumentError, "invalid limit: 0 for each_line"
-    end
-    unless block_given?
-      return to_enum(:__each_line, args, chomp) { nil }
-    end
-    __each_line(args, chomp) { |line| yield line }
-  end
-  alias each each_line
-
-  def self.for_fd(fd, mode = nil, **opts)
-    new(fd, mode, **opts)
-  end
-
-  # CRuby's rb_io_s_open: `new(*args)` plus block handling — the IO is
-  # closed when the block exits (via #close dispatch, so overrides run);
-  # an IOError meaning "already closed" is swallowed, everything else
-  # (StandardError or not) propagates.
-  def self.open(*args, **opts)
-    io = new(*args, **opts)
-    return io unless block_given?
-    begin
-      yield io
-    ensure
-      begin
-        io.close
-      rescue IOError => e
-        raise unless e.message.include?("closed stream")
-      end
-    end
-  end
-end
-
-class FloatDomainError < RangeError; end
-
-class Encoding
-  class CompatibilityError < EncodingError; end
-  class InvalidByteSequenceError < EncodingError; end
-  class UndefinedConversionError < EncodingError; end
-  class ConverterNotFoundError < EncodingError; end
-  class Converter; end
-
-  def self.default_internal
-    $DEFAULT_INTERNAL
-  end
-end
+# ---------------------------------------------------------------------------
+# Pure-Ruby builtins, one file per class/module. Load order is significant:
+#  * everything up to `comparable` must not reference Comparable/Enumerable
+#    at load time (e.g. ObjectSpace::WeakMap deliberately skips its
+#    `include ::Enumerable` because the module is not defined yet);
+#  * `io_buffer` (IO::Buffer includes Comparable) loads after `comparable`;
+#  * the class files after `enumerable` may `include Enumerable`;
+#  * the spec-compat alias fixups inside each class file run when that file
+#    loads, after the natives and Ruby methods they re-point.
+# ---------------------------------------------------------------------------
+
+require_relative 'basic_object'
+require_relative 'kernel'
+require_relative 'object'
+require_relative 'class'
+require_relative 'regexp'
+require_relative 'exception'
+require_relative 'module'
+require_relative 'warning'
+require_relative 'process'
+require_relative 'file'
+require_relative 'fiber'
+require_relative 'thread'
+require_relative 'boolean'
+require_relative 'marshal'
+require_relative 'gc'
+require_relative 'object_space'
+require_relative 'io'
+require_relative 'encoding'
 
 # TOPLEVEL_BINDING is defined by the runtime (Executor::init): a Binding
 # over an empty, outer-less frame that the *main script* is then executed
@@ -2254,182 +215,8 @@ end
 # locals — not this file's.
 
 require_relative 'comparable'
-
-class IO
-  class Buffer
-    include Comparable
-
-    # Error hierarchy (CRuby io_buffer.c). The Buffer class itself is
-    # native; only the exception constants live here.
-    class AllocationError < RuntimeError; end
-    class AccessError < RuntimeError; end
-    class LockedError < RuntimeError; end
-    class InvalidatedError < RuntimeError; end
-    class MaskError < ArgumentError; end
-  end
-end
-
-# Data class (Ruby 3.2+): immutable value objects with a keyword-based
-# initializer. Built on top of Struct for storage, attribute readers,
-# equality, hashing and inspection; the keyword initializer, argument
-# validation, `new`/`[]` positional-to-keyword coercion, frozen instances
-# and `with` are layered on here.
-class Data
-  def self.define(*members, &block)
-    members = members.map do |m|
-      case m
-      when Symbol then m
-      when String then m.to_sym
-      else raise TypeError, "#{m} is not a symbol"
-      end
-    end
-    klass = ::Struct.new(*members)
-    klass.class_eval do
-      # CRuby renders `Data` instances as `#<data ...>` rather than
-      # `#<struct ...>`. Reuse Struct's (correct) inspect — which already
-      # handles qualified/anonymous class names and bypasses an overridden
-      # `#name` — and only relabel the prefix.
-      def inspect
-        ::Struct.instance_method(:inspect).bind(self).call.sub(/\A#<struct/, '#<data')
-      end
-      alias_method :to_s, :inspect
-
-      # The keyword initializer is layered in a module so a user-defined
-      # `initialize` (from the `define` block) can `super` into it.
-      include(::Module.new do
-        def initialize(*args, **kw)
-          ms = self.class.members
-          kw = ::Hash[ms.zip(args)].merge(kw) unless args.empty?
-          values = ::Data.__data_values(ms, kw)
-          super(*values)
-          freeze
-        end
-      end)
-
-      # `new` / `[]` accept positional *or* keyword arguments; positional
-      # ones are zipped onto the members, then `initialize` (possibly
-      # user-overridden) is dispatched with keywords.
-      def self.new(*args, **kw)
-        ::Data.__data_alloc_init(self, args, kw)
-      end
-      class << self
-        alias_method :[], :new
-      end
-
-      # Returns a frozen copy with the given members replaced. Allocates
-      # and initializes directly rather than going through `new`, matching
-      # CRuby (a redefined `new` must not affect `with`).
-      def with(**kw)
-        return self if kw.empty?
-        norm = {}
-        kw.each { |k, v| norm[k.is_a?(::String) ? k.to_sym : k] = v }
-        copy = self.class.allocate
-        copy.send(:initialize, **to_h.merge(norm))
-        copy
-      end
-
-      # `deconstruct_keys(keys)` for pattern matching: `nil` returns all
-      # members; otherwise the requested keys (Symbol / String / `#to_str`)
-      # are looked up, stopping at the first non-member.
-      def deconstruct_keys(keys)
-        return to_h if keys.nil?
-        unless keys.is_a?(::Array)
-          raise TypeError, "wrong argument type #{keys.class} (expected Array or nil)"
-        end
-        ms = self.class.members
-        return {} if keys.size > ms.size
-        result = {}
-        keys.each do |k|
-          sym, rkey =
-            case k
-            when ::Symbol then [k, k]
-            when ::String then [k.to_sym, k]
-            else
-              if k.respond_to?(:to_str)
-                s = k.to_str
-                unless s.is_a?(::String)
-                  raise TypeError, "can't convert #{k.class} into String"
-                end
-                [s.to_sym, s]
-              else
-                raise TypeError, "#{k} is not a symbol nor a string"
-              end
-            end
-          break unless ms.include?(sym)
-          result[rkey] = send(sym)
-        end
-        result
-      end
-    end
-    klass.class_eval(&block) if block
-    klass
-  end
-
-  # The base initializer, reachable as `Data.instance_method(:initialize)`
-  # (used by e.g. marshalling libraries to populate an allocated instance).
-  def initialize(**kw)
-    ms = self.class.members
-    values = ::Data.__data_values(ms, kw)
-    ms.each_with_index { |m, i| send("#{m}=", values[i]) }
-    freeze
-  end
-
-  def self.__data_alloc_init(klass, args, kw)
-    ms = klass.members
-    unless args.empty?
-      raise ArgumentError, "wrong number of arguments (given #{args.size}, expected 0)" unless kw.empty?
-      unless args.size == ms.size
-        raise ArgumentError, "wrong number of arguments (given #{args.size}, expected 0)"
-      end
-      kw = ::Hash[ms.zip(args)]
-    end
-    obj = klass.allocate
-    obj.send(:initialize, **kw)
-    obj
-  end
-
-  # Validate `kw` against `members` (converting String / `#to_str` keys to
-  # Symbols) and return the member values in declaration order.
-  def self.__data_values(members, kw)
-    norm = {}
-    unknown = []
-    kw.each do |k, v|
-      key, disp = __data_key(k)
-      if members.include?(key)
-        norm[key] = v
-      else
-        unknown << disp
-      end
-    end
-    missing = members.reject { |m| norm.key?(m) }
-    unless missing.empty?
-      s = missing.size == 1 ? "" : "s"
-      raise ArgumentError, "missing keyword#{s}: #{missing.map { |m| ":#{m}" }.join(", ")}"
-    end
-    unless unknown.empty?
-      s = unknown.size == 1 ? "" : "s"
-      raise ArgumentError, "unknown keyword#{s}: #{unknown.join(", ")}"
-    end
-    members.map { |m| norm[m] }
-  end
-
-  # Normalize a keyword key to `[symbol, display]`, where `display` is how
-  # the key appears in an "unknown keyword" message (`:sym` / `"str"`).
-  def self.__data_key(k)
-    case k
-    when Symbol then [k, ":#{k}"]
-    when String then [k.to_sym, k.inspect]
-    else
-      if k.respond_to?(:to_str)
-        s = k.to_str
-        raise TypeError, "can't convert #{k.class} into String" unless s.is_a?(String)
-        [s.to_sym, s.inspect]
-      else
-        raise TypeError, "#{k} is not a symbol nor a string"
-      end
-    end
-  end
-end unless defined?(::Data)
+require_relative 'io_buffer'
+require_relative 'data'
 
 require_relative 'enumerable'
 IO.include(Enumerable)
@@ -2446,209 +233,17 @@ require_relative 'symbol'
 require_relative 'error'
 require_relative 'set'
 require_relative 'struct'
-require_relative 'builtins'
+require_relative 'monitor'
+require_relative 'enumerator'
+require_relative 'hash'
+require_relative 'file_stat'
+require_relative 'proc'
+require_relative 'method'
+require_relative 'dir'
+require_relative 'time'
+require_relative 'match_data'
+require_relative 'filetest'
+require_relative 'env'
 require_relative 'pathname_builtins'
 
-# Minimal ARGF implementation. Full semantics (transparent line-by-line
-# reading across ARGV files with $_/$.) are only partially implemented;
-# enough shape is provided for specs to load and for simple cases to
-# proceed. Deliberately defined after enumerable is loaded so the
-# `include Enumerable` on the class body resolves.
-class ARGFClass
-  include Enumerable
-
-  def initialize(*argv)
-    @argv = argv.empty? ? (defined?(::ARGV) ? ::ARGV : []) : argv
-    @current_file = nil
-    @current_name = nil
-    @lineno = 0
-    # Set once every input stream has been consumed. ARGF must NOT fall
-    # back to $stdin after named ARGV files are exhausted (CRuby reads
-    # stdin only when ARGV was empty to begin with) — doing so made
-    # `ARGF.read; ARGF.readlines` block forever on stdin.
-    @exhausted = false
-  end
-
-  def argv
-    @argv
-  end
-
-  def filename
-    @current_name || (@argv.first || '-')
-  end
-  alias_method :path, :filename
-
-  def file
-    __stream || @current_file || $stdin
-  end
-
-  # The current open stream, opening the next ARGV file (or binding
-  # $stdin when ARGV was empty from the start) as needed. Returns nil
-  # once every input has been consumed — never falling back to $stdin
-  # after named files are done.
-  def __stream
-    return @current_file if @current_file && !@current_file.closed?
-    return nil if @exhausted
-    if @argv.empty?
-      if @current_name
-        # Named files were consumed earlier; there is no more input.
-        @exhausted = true
-        return nil
-      end
-      @current_name = '-'
-      @current_file = $stdin
-    else
-      @current_name = @argv.shift
-      @current_file = @current_name == '-' ? $stdin : File.open(@current_name)
-    end
-    @current_file
-  end
-  private :__stream
-
-  # Close out the current stream (it hit EOF) and mark ARGF exhausted
-  # when no further ARGV entries remain.
-  def __finish_stream
-    f = @current_file
-    f.close if f && !f.closed? && !f.equal?($stdin)
-    @current_file = nil
-    @exhausted = true if @argv.empty?
-  end
-  private :__finish_stream
-
-  def advance
-    !!__stream
-  end
-
-  def lineno;     @lineno; end
-  def lineno=(n); @lineno = n; end
-  def pos;        0;       end
-  alias_method :tell, :pos
-  def pos=(n);    n;       end
-  def closed?
-    !@current_file || @current_file.closed?
-  end
-  def close
-    @current_file.close if @current_file && !@current_file.closed? && @current_file != $stdin
-    self
-  end
-  def eof?
-    @argv.empty? && (@current_file.nil? || @current_file.closed?)
-  end
-  alias_method :eof, :eof?
-  def skip
-    if @current_file && !@current_file.closed? && @current_file != $stdin
-      @current_file.close
-    end
-    @current_file = nil
-    self
-  end
-  def rewind
-    @current_file.rewind if @current_file && !@current_file.closed?
-    @lineno = 0
-  end
-  def each
-    return to_enum(:each) unless block_given?
-    while (f = __stream)
-      f.each_line do |line|
-        @lineno += 1
-        yield line
-      end
-      __finish_stream
-    end
-    self
-  end
-  alias_method :each_line, :each
-  alias_method :lines, :each
-  def readlines(*args)
-    result = []
-    each { |line| result << line }
-    result
-  end
-  # CRuby-compatible ARGF.read (verified against CRuby 4.0.2):
-  # - read           -> the concatenation of every remaining stream; ""
-  #                     for empty input; nil once ARGF is exhausted.
-  # - read(len)      -> up to len bytes, continuing across file
-  #                     boundaries; stops at exactly len (this bound is
-  #                     what keeps `ARGF.read(100)` on /dev/zero from
-  #                     reading forever); nil at EOF; "" for len == 0.
-  # - read(len, buf) -> fills and returns buf (same object).
-  def read(length = nil, outbuf = nil)
-    if length && length < 0
-      raise ArgumentError, "negative length #{length} given"
-    end
-    buf = String.new
-    had_stream = false
-    read_any = false
-    unless length == 0
-      loop do
-        break if length && buf.bytesize >= length
-        f = __stream
-        break unless f
-        had_stream = true
-        need = length && length - buf.bytesize
-        chunk = need ? f.read(need) : f.read
-        if chunk && !chunk.empty?
-          read_any = true
-          buf << chunk
-        end
-        # A short (or nil) chunk means this stream hit EOF; a full read
-        # without length always consumes the stream.
-        if need.nil? || chunk.nil? || chunk.bytesize < need
-          __finish_stream
-        end
-      end
-    end
-    result =
-      if length.nil?
-        had_stream ? buf : nil
-      elsif length == 0
-        ""
-      else
-        read_any ? buf : nil
-      end
-    if outbuf
-      outbuf.replace(result || "")
-      result = outbuf if result
-    end
-    result
-  end
-  def readline(*args)
-    line = nil
-    each { |l| line = l; break }
-    raise EOFError, "end of file reached" if line.nil?
-    line
-  end
-  def gets(*args)
-    line = nil
-    each { |l| line = l; break }
-    line
-  end
-  def getc;     nil; end
-  def readchar; raise EOFError; end
-  def getbyte;  nil; end
-  def readbyte; raise EOFError; end
-  def inspect;  'ARGF'; end
-  def to_s;     'ARGF'; end
-  def to_a
-    readlines
-  end
-  def fileno
-    file.fileno
-  end
-  alias_method :to_i, :fileno
-  def to_io
-    file
-  end
-  def binmode;  self; end
-  def binmode?; false; end
-  def external_encoding; Encoding.default_external; end
-  def internal_encoding; Encoding.default_internal; end
-  def inplace_mode;  nil; end
-  def inplace_mode=(v); v; end
-  def set_encoding(*); self; end
-  def write(*args); args.map(&:to_s).join.bytesize; end
-  def print(*args); $stdout.print(*args); end
-  def puts(*args);  $stdout.puts(*args); end
-end
-
-ARGF = ARGFClass.new
+require_relative 'argf'
