+++ monoruby/builtins/struct.rb
@@ -182,3 +182,12 @@ class Struct
   end
   alias filter select
 end
+
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Struct
+  alias inspect to_s
+end
