+++ monoruby/builtins/symbol.rb
@@ -72,3 +72,15 @@ class Symbol
     self.to_s.match?(other, pos)
   end
 end
+
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Symbol
+  alias === ==
+  alias id2name to_s
+  alias intern to_sym
+  alias size length
+end
