+++ monoruby/builtins/time.rb
@@ -1,0 +1,9 @@
+# True alias required by ruby/spec's strict
+# `Klass.instance_method(:a) == Klass.instance_method(:b)` identity checks
+# (the 2026-06 "rely less on shared examples" refactoring wave). The
+# aliased original must already be defined (natively or above) by now.
+
+class Time
+  include Comparable
+  alias asctime ctime
+end
